backend/pyproject.toml
[tool.poetry]
name = "FastAPI Server"
version = "0.1.0"
description = ""
authors = ["Nikolay <nchernov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
requests = "^2.28.1"
fastapi-pagination = "^0.11.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
pydantic = "^1.10.4"
python-dotenv = "^0.21.0"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"
sqlalchemy = "^1.4.45"
email-validator = "^1.3.0"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.1"
invoke = "^1.7.3"
ipython = "^8.7.0"
sqladmin = "^0.8.0"
anyio = "^3.6.2"
setuptools = "^65.6.3"
phonenumbers = "^8.13.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
pytest = "^7.1.3"
autopep8 = "^1.7.0"
pytest-asyncio = "^0.20.1"
pytest-mock = "^3.10.0"
greenlet = "^2.0.1"
faker = "^15.3.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
